remove now unecc EmptySSLCertificate entity, factor helper methods into seperate extension files
